Feature to set Offline flag for ff client
Why: To optimize the ff client

How: Added a new function named SetOffline, and that takes boolean as a
parameter, and returns the updated value. It will start or stop the flag
updater daemon based on the input. The background_updater is also
updated to check for close() signals.

Tags: [ffclient, enhancement, feature]