feat(lib): set Offline flag for ff client (#1774)
* Feature to set Offline flag for ff client

Why: To optimize the ff client

How: Added a new function named SetOffline, and that takes boolean as a
parameter, and returns the updated value. It will start or stop the flag
updater daemon based on the input. The background_updater is also
updated to check for close() signals.

Tags: [ffclient, enhancement, feature]

* Added sync.Once

* fix: race condition

Signed-off-by: Thomas Poignant <thomas.poignant@gofeatureflag.org>

* fix data race

Signed-off-by: Thomas Poignant <thomas.poignant@gofeatureflag.org>

* remove unused checks

Signed-off-by: Thomas Poignant <thomas.poignant@gofeatureflag.org>

---------

Signed-off-by: Thomas Poignant <thomas.poignant@gofeatureflag.org>
Co-authored-by: Thomas Poignant <thomas.poignant@gofeatureflag.org>