Fixes a potential difficult-to-diagnose failure
When we deployed a service called debug_test_service, EKS
generated an env variable called DEBUG_TEST_SERVICE which
goff understood to enable debug, but the value
of the service could not be coerced into boolean
so the goff service started crash looping.

This failure caused us a 2 hour headache.